fix: style코드 분리
