+++ src/components/Calendar.tsx
@@ -4,68 +4,13 @@ import { styled } from '@mui/material/styles';
 import dayjs, { Dayjs } from 'dayjs';
 import React, { useEffect, useState } from 'react';
 
-export const CalanderDay = styled('div')({
-  display: 'flex',
-  justifyContent: 'center',
-  alignItems: 'center',
-  width: '36px',
-  height: '36px',
-  textAlign: 'center',
-  lineHeight: '16px',
-});
-
-export const CalanderDayBtn = styled('button')<{ today?: boolean; isSelected?: boolean }>`
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  width: 36px;
-  height: 36px;
-  border: 0;
-  background: 0;
-  padding: 0;
-  position: relative;
-
-  text-align: center;
-  font-size: 16px;
-  font-weight: 400;
-  line-height: 20px;
-
-  &::after {
-    content: '오늘';
-    display: ${(props) => (props.today ? 'flex' : 'none')};
-    position: absolute;
-    font-size: 10px;
-    font-weight: 400;
-    line-height: 12px;
-    bottom: -14px;
-    color: ${(props) => props.theme.palette.text.primary};
-  }
-  ${(props) =>
-    props.isSelected && {
-      color: 'white',
-      fontWeight: 600,
-      backgroundColor: props.theme.palette.primary.main,
-      borderRadius: '20%',
-    }}
-`;
-
-export const CalanderWeek = styled('div')({
-  display: 'flex',
-  justifyContent: 'space-between',
-  alignItems: 'center',
-  paddingBottom: '16px',
-});
-
-const CalanderContent = styled('div')({
-  display: 'flex',
-  flexDirection: 'column',
-  flex: 1,
-});
-const CalenderWrapper = styled('div')({
-  padding: '0px 0px 32px',
-  flex: 1,
-  width: '100%',
-});
+import {
+  CalanderContent,
+  CalanderDay,
+  CalanderDayBtn,
+  CalanderWeek,
+  CalenderWrapper,
+} from './styled';
 
 const RenderDays = () => {
   const days = [];
